version 0.5 with code and test fixes, adding readme

+++ src/main/java/com/ione/controller/CustomerController.java
@@ -2,9 +2,11 @@ package com.ione.controller;
 
 import com.ione.dto.CustomerRequestDTO;
 import com.ione.dto.CustomerResponseDTO;
+import com.ione.security.AuthUtil;
 import com.ione.service.CustomerService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
@@ -36,12 +38,21 @@ public class CustomerController {
     @PreAuthorize("hasRole('CUSTOMER')")
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteCustomer(@PathVariable Integer id) {
+        String email = AuthUtil.getCurrentUsername();
+        if (customerService.isNotOwner(email, id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         customerService.deleteCustomer(id);
         return ResponseEntity.noContent().build();
     }
+
     @PreAuthorize("hasRole('CUSTOMER')")
-    @PostMapping("/{id}/balance")
+    @PatchMapping("/{id}/balance")
     public ResponseEntity<String> updateBalance(@PathVariable Integer id, @RequestParam BigDecimal amount) {
+        String email = AuthUtil.getCurrentUsername();
+        if (customerService.isNotOwner(email, id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         customerService.updateBalance(id, amount);
         return ResponseEntity.ok("Customer balance updated");
     }
