+++ src/main/java/com/ione/controller/DriverController.java
@@ -2,9 +2,11 @@ package com.ione.controller;
 
 import com.ione.dto.DriverRequestDTO;
 import com.ione.dto.DriverResponseDTO;
+import com.ione.security.AuthUtil;
 import com.ione.service.DriverService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
@@ -36,14 +38,23 @@ public class DriverController {
         return ResponseEntity.ok(driverService.getAllDrivers());
     }
 
+    @PreAuthorize("hasRole('DRIVER')")
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteDriver(@PathVariable Integer id) {
+        String email = AuthUtil.getCurrentUsername();
+        if (driverService.isNotOwner(email, id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         driverService.deleteDriver(id);
         return ResponseEntity.noContent().build();
     }
     @PreAuthorize("hasRole('DRIVER')")
-    @PostMapping("/{id}/balance")
+    @PatchMapping("/{id}/balance")
     public ResponseEntity<String> updateBalance(@PathVariable Integer id, @RequestParam BigDecimal amount) {
+        String email = AuthUtil.getCurrentUsername();
+        if (driverService.isNotOwner(email, id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         driverService.updateBalance(id, amount);
         return ResponseEntity.ok("Driver balance updated");
     }
